Replaced a few toInt() or roundToInt() references with crash-safe ones; removed UByte,UShort,UInt,ULong to discourage using them (because they just duplicate so much code for one extra bit)

+++ KOML/src/org/joml/Vector2i.kt
@@ -73,7 +73,7 @@ open class Vector2i(
     fun distance(x: Int, y: Int): Double {
         val dx = this.x - x
         val dy = this.y - y
-        return sqrt((dx.toLong() * dx + dy.toLong() * dy).toFloat()).toDouble()
+        return sqrt((dx.toLong() * dx + dy.toLong() * dy).toDouble())
     }
 
     fun distanceSquared(v: Vector2i) = distanceSquared(v.x, v.y)
